LayoutTests/inspector/extensions-events.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test2.js"></script>
<script src="extensions-test.js"></script>
<script type="text/javascript">


function extension_testOnSelectionChanged(nextTest)
{
    function onSelectionChanged()
    {
        webInspector.panels.elements.onSelectionChanged.removeListener(onSelectionChanged);
        output("onSelectionChanged fired");
        nextTest();
    }
    webInspector.panels.elements.onSelectionChanged.addListener(onSelectionChanged);
    webInspector.inspectedWindow.eval("inspect(document.body.children[0])");
}

function extension_testOnResourceFinished(nextTest)
{
    function onFinished()
    {
        webInspector.resources.onFinished.removeListener(onFinished);
        output("onFinished fired");
        nextTest();
    }
    webInspector.resources.onFinished.addListener(onFinished);
    webInspector.inspectedWindow.eval("var xhr = new XMLHttpRequest(); xhr.open('GET', '" + location.href + "', false); xhr.send(null);");
}

function extension_testPageLoadEvents(nextTest)
{
    var callbackCount = 0;
    webInspector.inspectedWindow.onDOMContentLoaded.addListener(function(time) {
        output("onDOMContentLoaded fired: " + typeof time);
    });
    webInspector.inspectedWindow.onLoaded.addListener(function(time) {
        output("onLoaded fired: " + typeof time);
        if (++callbackCount > 1)
            nextTest();
    });
    dispatchOnFrontend({ command: "reload" }, function() {
        if (++callbackCount > 1)
            nextTest();
    });
}

</script>
</head>
<body onload="runTest()">
<p>Tests WebInspector extension API</p>
</body>
</html>
